requirements.txt
accelerate==1.9.0
    # via -r requirements.in
aiodns==3.5.0
    # via ccxt
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.14
    # via ccxt
aiolimiter==1.2.1
    # via python-telegram-bot
aiosignal==1.4.0
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   openai
apscheduler==3.11.0
    # via
    #   -r requirements.in
    #   python-telegram-bot
attrs==25.3.0
    # via aiohttp
blinker==1.9.0
    # via flask
cachetools==6.1.0
    # via python-telegram-bot
ccxt==4.4.95
    # via -r requirements.in
certifi==2025.7.14
    # via
    #   ccxt
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via
    #   cryptography
    #   pycares
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via flask
colorama==0.4.6
    # via
    #   click
    #   loguru
    #   tqdm
cryptography==45.0.5
    # via ccxt
dash==3.1.1
    # via -r requirements.in
diskcache==5.6.3
    # via llama-cpp-python
distro==1.9.0
    # via openai
filelock==3.18.0
    # via
    #   huggingface-hub
    #   torch
    #   transformers
flask==3.1.1
    # via dash
frozenlist==1.7.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2025.7.0
    # via
    #   huggingface-hub
    #   torch
greenlet==3.2.3
    # via sqlalchemy
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   openai
    #   python-telegram-bot
huggingface-hub==0.33.4
    # via
    #   accelerate
    #   tokenizers
    #   transformers
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
importlib-metadata==8.7.0
    # via dash
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via
    #   flask
    #   llama-cpp-python
    #   torch
jiter==0.10.0
    # via openai
joblib==1.5.1
    # via scikit-learn
llama-cpp-python==0.3.13
    # via -r requirements.in
loguru==0.7.3
    # via -r requirements.in
markupsafe==3.0.2
    # via
    #   flask
    #   jinja2
    #   werkzeug
mpmath==1.3.0
    # via sympy
multidict==6.6.3
    # via
    #   aiohttp
    #   yarl
narwhals==1.47.1
    # via plotly
nest-asyncio==1.6.0
    # via dash
networkx==3.4
    # via torch
numpy==2.2.6
    # via
    #   -r requirements.in
    #   accelerate
    #   llama-cpp-python
    #   pandas
    #   scikit-learn
    #   scipy
    #   ta
    #   torchvision
    #   transformers
    #   xgboost
openai==1.97.0
    # via -r requirements.in
packaging==25.0
    # via
    #   accelerate
    #   huggingface-hub
    #   plotly
    #   transformers
pandas==2.3.1
    # via
    #   -r requirements.in
    #   ta
pillow==11.3.0
    # via torchvision
plotly==6.2.0
    # via
    #   -r requirements.in
    #   dash
propcache==0.3.2
    # via
    #   aiohttp
    #   yarl
psutil==7.0.0
    # via accelerate
pycares==4.9.0
    # via aiodns
pycparser==2.22
    # via cffi
pydantic==2.11.7
    # via openai
pydantic-core==2.33.2
    # via pydantic
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.1.1
    # via -r requirements.in
python-telegram-bot[ext]==22.2
    # via -r requirements.in
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via
    #   accelerate
    #   huggingface-hub
    #   transformers
regex==2024.11.6
    # via transformers
requests==2.32.4
    # via
    #   -r requirements.in
    #   ccxt
    #   dash
    #   huggingface-hub
    #   transformers
retrying==1.4.0
    # via
    #   -r requirements.in
    #   dash
safetensors==0.5.3
    # via
    #   accelerate
    #   transformers
scikit-learn==1.7.1
    # via -r requirements.in
scipy==1.15.3
    # via
    #   scikit-learn
    #   xgboost
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   openai
sqlalchemy==2.0.41
    # via -r requirements.in
sympy==1.14.0
    # via torch
ta==0.11.0
    # via -r requirements.in
threadpoolctl==3.6.0
    # via scikit-learn
tokenizers==0.21.2
    # via transformers
torch==2.7.1
    # via
    #   -r requirements.in
    #   accelerate
    #   torchaudio
    #   torchvision
torchaudio==2.7.1
    # via -r requirements.in
torchvision==0.22.1
    # via -r requirements.in
tornado==6.5.1
    # via python-telegram-bot
tqdm==4.67.1
    # via
    #   huggingface-hub
    #   openai
    #   transformers
transformers==4.53.2
    # via -r requirements.in
typing-extensions==4.14.1
    # via
    #   aiosignal
    #   anyio
    #   ccxt
    #   dash
    #   huggingface-hub
    #   llama-cpp-python
    #   openai
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   torch
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
tzdata==2025.2
    # via
    #   pandas
    #   tzlocal
tzlocal==5.3.1
    # via apscheduler
urllib3==2.5.0
    # via requests
werkzeug==3.1.3
    # via
    #   dash
    #   flask
win32-setctime==1.2.0
    # via loguru
xgboost==3.0.2
    # via -r requirements.in
yarl==1.20.1
    # via
    #   aiohttp
    #   ccxt
zipp==3.23.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
